Add aliases, DummyModify, doc Resolver, fix docs
Add aliases
-----------

The `alias` method on `TransformedTree` was a `todo!`. It is now
implemented (but not tested)

We also add a `alias_mods` that directly inserts fully formed modifiers
rather than text.

DummyModify
-----------

A `Modify` implementation that does nothing. Useful for test code

doc Resolver
------------

Document the `Resolver` trait, detailing what it does and what are the
existing implementations.

fix docs
--------

`cargo doc` was generating a lot of warnings, we address all of those
warnings.